[drc] -Fix: missing cast caused overflow on circle shape hole ring break check, missing legit broken rings
git-svn-id: svn://repo.hu/pcb-rnd/trunk@24023 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/obj_pstk.c
@@ -1070,7 +1070,7 @@ static pcb_bool pcb_pstk_shape_hole_break(pcb_pstk_shape_t *shp, pcb_coord_t hdi
 		case PCB_PSSH_HSHADOW:
 			break;
 		case PCB_PSSH_CIRC:
-			dist = sqrt(shp->data.circ.x*shp->data.circ.x + shp->data.circ.y*shp->data.circ.y);
+			dist = sqrt((double)shp->data.circ.x*(double)shp->data.circ.x + (double)shp->data.circ.y*(double)shp->data.circ.y);
 			neck = (double)(shp->data.circ.dia - hdia) / 2.0 - dist;
 			break;
 		case PCB_PSSH_LINE:
